Prmts 0.0.6

+++ thesisbackend/src/main/java/com/thesis/backend/model/dto/CompartmentDto.java
@@ -3,20 +3,18 @@ package com.thesis.backend.model.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Column;
-
 @Getter
 @Setter
 public class CompartmentDto {
     private int id;
     private String name;
-    private int xDimension;
-    private int yDimension;
-    private int width;
-    private int depth;
-    private int xKonva;
-    private int yKonva;
-    private int widthKonva;
-    private int heightKonva;
+    private float xDimension;
+    private float yDimension;
+    private float width;
+    private float depth;
+    private float xKonva;
+    private float yKonva;
+    private float widthKonva;
+    private float heightKonva;
     private int floorId;
 }
